+++ examples/resources/iosxr_interface/resource.tf
@@ -1,8 +1,21 @@
 resource "iosxr_interface" "example" {
-  interface_name          = "GigabitEthernet0/0/0/1"
-  l2transport             = false
-  point_to_point          = false
-  multipoint              = false
+  interface_name                  = "GigabitEthernet0/0/0/1"
+  l2transport                     = false
+  point_to_point                  = false
+  multipoint                      = false
+  dampening_decay_half_life_value = 2
+  ipv4_point_to_point             = true
+  input_policy = [
+    {
+      service_policy_name = "CORE-INPUT-POLICY"
+    }
+  ]
+  output_policy = [
+    {
+      service_policy_name = "CORE-OUTPUT-POLICY"
+    }
+  ]
+  bfd_mode_ietf           = true
   shutdown                = true
   mtu                     = 9000
   bandwidth               = 100000