+++ internal/provider/data_source_iosxr_router_isis_address_family_test.go
@@ -35,6 +35,7 @@ func TestAccDataSourceIosxrRouterISISAddressFamily(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "microloop_avoidance_protected", "false"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "microloop_avoidance_segment_routing", "true"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "advertise_passive_only", "true"),
+					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "advertise_link_attributes", "true"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "mpls_ldp_auto_config", "false"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "mpls_traffic_eng_router_id_ip_address", "1.2.3.4"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "mpls_traffic_eng_level_1_2", "false"),
@@ -46,6 +47,8 @@ func TestAccDataSourceIosxrRouterISISAddressFamily(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "spf_prefix_priorities.0.priority", "critical"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "spf_prefix_priorities.0.tag", "100"),
 					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "segment_routing_mpls_sr_prefer", "true"),
+					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "maximum_redistributed_prefixes", "100"),
+					resource.TestCheckResourceAttr("data.iosxr_router_isis_address_family.test", "maximum_redistributed_prefixes_levels.0.level_id", "1"),
 				),
 			},
 		},
@@ -79,6 +82,7 @@ resource "iosxr_router_isis_address_family" "test" {
 	microloop_avoidance_protected = false
 	microloop_avoidance_segment_routing = true
 	advertise_passive_only = true
+	advertise_link_attributes = true
 	mpls_ldp_auto_config = false
 	mpls_traffic_eng_router_id_ip_address = "1.2.3.4"
 	mpls_traffic_eng_level_1_2 = false
@@ -92,6 +96,10 @@ resource "iosxr_router_isis_address_family" "test" {
 		tag = 100
 	}]
 	segment_routing_mpls_sr_prefer = true
+	maximum_redistributed_prefixes = 100
+	maximum_redistributed_prefixes_levels = [{
+		level_id = 1
+	}]
 }
 
 data "iosxr_router_isis_address_family" "test" {