+++ internal/provider/resource_iosxr_router_isis_interface_address_family_test.go
@@ -21,6 +21,7 @@ func TestAccIosxrRouterISISInterfaceAddressFamily(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxr_router_isis_interface_address_family.test", "fast_reroute_per_prefix_levels.0.level_id", "1"),
 					resource.TestCheckResourceAttr("iosxr_router_isis_interface_address_family.test", "fast_reroute_per_prefix_levels.0.ti_lfa", "true"),
 					resource.TestCheckResourceAttr("iosxr_router_isis_interface_address_family.test", "tag", "100"),
+					resource.TestCheckResourceAttr("iosxr_router_isis_interface_address_family.test", "advertise_prefix_route_policy", "ROUTE_POLICY_1"),
 				),
 			},
 			{
@@ -55,6 +56,7 @@ func testAccIosxrRouterISISInterfaceAddressFamilyConfig_all() string {
 			ti_lfa = true
 		}]
 		tag = 100
+		advertise_prefix_route_policy = "ROUTE_POLICY_1"
 	}
 	`
 }