+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Fix incompatibility with gNMI and IOS-XR 7.6
 - Remove `iosxr_segment_routing` resource and data source due to unified model being deprecated
+- Add `iosxr_router_bgp_vrf_neighbor_address_family` resource and data source
+- Add `iosxr_router_bgp_neighbor_group_address_family` resource and data source
+- Add `iosxr_router_static` resource and data source
 
 ## 0.1.7